riscv_h::register::hcounteren

Struct Hcounteren

pub struct Hcounteren { /* private fields */ }

Hypervisor Counter Enable Register.

Implementations

impl Hcounteren

pub fn bits(&self) -> usize

Returns the raw bits of the register.

pub fn from_bits(x: usize) -> Self

Creates a register value from raw bits.

pub unsafe fn write(&self)

Writes the register value to the CSR.

pub fn cy(&self) -> bool

Returns the status of the cycle counter.

pub fn set_cy(&mut self, val: bool)

Sets the status of the cycle counter.

pub fn tm(&self) -> bool

Returns the status of the timer counter.

pub fn set_tm(&mut self, val: bool)

Sets the status of the timer counter.

pub fn ir(&self) -> bool

Returns the status of the interrupt counter.

pub fn set_ir(&mut self, val: bool)

Sets the status of the interrupt counter.

pub fn hpm3(&self) -> bool

Returns the status of the hypervisor performance monitor 3.

pub fn set_hpm3(&mut self, val: bool)

Sets the status of the hypervisor performance monitor 3.

pub fn hpm4(&self) -> bool

Returns the status of the hypervisor performance monitor 4.

pub fn set_hpm4(&mut self, val: bool)

Sets the status of the hypervisor performance monitor 4.

pub fn hpm5(&self) -> bool

Returns the status of the hypervisor performance monitor 5.

pub fn set_hpm5(&mut self, val: bool)

Sets the status of the hypervisor performance monitor 5.

pub fn hpm6(&self) -> bool

Returns the status of the hypervisor performance monitor 6.

pub fn set_hpm6(&mut self, val: bool)

Sets the status of the hypervisor performance monitor 6.

pub fn hpm7(&self) -> bool

Returns the status of the hypervisor performance monitor 7.

pub fn set_hpm7(&mut self, val: bool)

Sets the status of the hypervisor performance monitor 7.

pub fn hpm8(&self) -> bool

Returns the status of the hypervisor performance monitor 8.

pub fn set_hpm8(&mut self, val: bool)

Sets the status of the hypervisor performance monitor 8.

pub fn hpm9(&self) -> bool

Returns the status of the hypervisor performance monitor 9.

pub fn set_hpm9(&mut self, val: bool)

Sets the status of the hypervisor performance monitor 9.

pub fn hpm10(&self) -> bool

Returns the status of the hypervisor performance monitor 10.

pub fn set_hpm10(&mut self, val: bool)

Sets the status of the hypervisor performance monitor 10.

pub fn hpm11(&self) -> bool

Returns the status of the hypervisor performance monitor 11.

pub fn set_hpm11(&mut self, val: bool)

Sets the status of the hypervisor performance monitor 11.

pub fn hpm12(&self) -> bool

Returns the status of the hypervisor performance monitor 12.

pub fn set_hpm12(&mut self, val: bool)

Sets the status of the hypervisor performance monitor 12.

pub fn hpm13(&self) -> bool

Returns the status of the hypervisor performance monitor 13.

pub fn set_hpm13(&mut self, val: bool)

Sets the status of the hypervisor performance monitor 13.

pub fn hpm14(&self) -> bool

Returns the status of the hypervisor performance monitor 14.

pub fn set_hpm14(&mut self, val: bool)

Sets the status of the hypervisor performance monitor 14.

pub fn hpm15(&self) -> bool

Returns the status of the hypervisor performance monitor 15.

pub fn set_hpm15(&mut self, val: bool)

Sets the status of the hypervisor performance monitor 15.

pub fn hpm16(&self) -> bool

Returns the status of the hypervisor performance monitor 16.

pub fn set_hpm16(&mut self, val: bool)

Sets the status of the hypervisor performance monitor 16.

pub fn hpm17(&self) -> bool

Returns the status of the hypervisor performance monitor 17.

pub fn set_hpm17(&mut self, val: bool)

Sets the status of the hypervisor performance monitor 17.

pub fn hpm18(&self) -> bool

Returns the status of the hypervisor performance monitor 18.

pub fn set_hpm18(&mut self, val: bool)

Sets the status of the hypervisor performance monitor 18.

pub fn hpm19(&self) -> bool

Returns the status of the hypervisor performance monitor 19.

pub fn set_hpm19(&mut self, val: bool)

Sets the status of the hypervisor performance monitor 19.

pub fn hpm20(&self) -> bool

Returns the status of the hypervisor performance monitor 20.

pub fn set_hpm20(&mut self, val: bool)

Sets the status of the hypervisor performance monitor 20.

pub fn hpm21(&self) -> bool

Returns the status of the hypervisor performance monitor 21.

pub fn set_hpm21(&mut self, val: bool)

Sets the status of the hypervisor performance monitor 21.

pub fn hpm22(&self) -> bool

Returns the status of the hypervisor performance monitor 22.

pub fn set_hpm22(&mut self, val: bool)

Sets the status of the hypervisor performance monitor 22.

pub fn hpm23(&self) -> bool

Returns the status of the hypervisor performance monitor 23.

pub fn set_hpm23(&mut self, val: bool)

Sets the status of the hypervisor performance monitor 23.

pub fn hpm24(&self) -> bool

Returns the status of the hypervisor performance monitor 24.

pub fn set_hpm24(&mut self, val: bool)

Sets the status of the hypervisor performance monitor 24.

pub fn hpm25(&self) -> bool

Returns the status of the hypervisor performance monitor 25.

pub fn set_hpm25(&mut self, val: bool)

Sets the status of the hypervisor performance monitor 25.

pub fn hpm26(&self) -> bool

Returns the status of the hypervisor performance monitor 26.

pub fn set_hpm26(&mut self, val: bool)

Sets the status of the hypervisor performance monitor 26.

pub fn hpm27(&self) -> bool

Returns the status of the hypervisor performance monitor 27.

pub fn set_hpm27(&mut self, val: bool)

Sets the status of the hypervisor performance monitor 27.

pub fn hpm28(&self) -> bool

Returns the status of the hypervisor performance monitor 28.

pub fn set_hpm28(&mut self, val: bool)

Sets the status of the hypervisor performance monitor 28.

pub fn hpm29(&self) -> bool

Returns the status of the hypervisor performance monitor 29.

pub fn set_hpm29(&mut self, val: bool)

Sets the status of the hypervisor performance monitor 29.

pub fn hpm30(&self) -> bool

Returns the status of the hypervisor performance monitor 30.

pub fn set_hpm30(&mut self, val: bool)

Sets the status of the hypervisor performance monitor 30.

pub fn hpm31(&self) -> bool

Returns the status of the hypervisor performance monitor 31.

pub fn set_hpm31(&mut self, val: bool)

Sets the status of the hypervisor performance monitor 31.