riscv_h::register

Module vsstatus

Virtual supervisor status register Virtual Supervisor Status Register.

The vsstatus register contains status and control fields for the virtual supervisor mode. This register controls various aspects of virtual machine execution including privilege levels, memory management, and floating-point state.

Structs

Vsstatus

Virtual Supervisor Status Register

Enums

UxlValues

Hypervisor User XLEN values.

Functions

clear_mxr⚠

Make executable readable enable.

clear_sie⚠

Supervisor interrupt enable.

clear_spie⚠

Supervisor previous interrupt enable.

clear_spp⚠

Supervisor previous privilege enable.

clear_sum⚠

Supervisor user memory enable.

clear_ube⚠

User binary endianness enable.

read

Reads the CSR.

set_mxr⚠

Make executable readable enable.

set_sie⚠

Supervisor interrupt enable.

set_spie⚠

Supervisor previous interrupt enable.

set_spp⚠

Supervisor previous privilege enable.

set_sum⚠

Supervisor user memory enable.

set_ube⚠

User binary endianness enable.

try_read

Attempts to reads the CSR.