riscv_h::register::vsstatus

Struct Vsstatus

pub struct Vsstatus { /* private fields */ }

Virtual Supervisor Status Register

Implementations

impl Vsstatus

pub fn bits(&self) -> usize

Returns the raw bits of the register.

pub fn from_bits(x: usize) -> Self

Creates a register value from raw bits.

pub unsafe fn write(&self)

Writes the register value to the CSR.

pub fn sd(&self) -> usize

Returns the status of the dirty state fields.

pub fn set_sd(&mut self, val: usize)

Sets the status of the dirty state fields.

pub fn uxl(&self) -> UxlValues

Returns the effective user XLEN setting.

pub fn set_uxl(&mut self, val: UxlValues)

Sets the effective user XLEN setting.

pub fn mxr(&self) -> bool

Returns the status of the make executable readable bit.

pub fn set_mxr(&mut self, val: bool)

Sets the MXR (Make eXecutable Readable) bit.

pub fn sum(&self) -> bool

Returns the status of the supervisor user memory access bit.

pub fn set_sum(&mut self, val: bool)

Sets the status of the supervisor user memory access bit.

pub fn xs(&self) -> usize

Returns the status of the extension state fields.

pub fn set_xs(&mut self, val: usize)

Sets the status of the extension state fields.

pub fn fs(&self) -> usize

Returns the floating point state.

pub fn set_fs(&mut self, val: usize)

Sets the floating point state.

pub fn spp(&self) -> bool

Returns the supervisor previous privilege.

pub fn set_spp(&mut self, val: bool)

Sets the supervisor previous privilege.

pub fn ube(&self) -> bool

Returns the user binary endianness.

pub fn set_ube(&mut self, val: bool)

Sets the user binary endianness.

pub fn spie(&self) -> bool

Returns the supervisor previous interrupt enable.

pub fn set_spie(&mut self, val: bool)

Sets the supervisor previous interrupt enable.

pub fn sie(&self) -> bool

Returns the supervisor interrupt enable.

pub fn set_sie(&mut self, val: bool)

Sets the supervisor interrupt enable.